Updated to include the possibility of putting code in the iguana directory that is mapped to org/iguanatool/ext

+++ setup.sh
@@ -7,10 +7,27 @@ if [ -z "$IGUANA_CASESTUDIES_HOME" ]; then
     exit 1
 fi
 
+# Make the case study C code symlink
 cd $IGUANA_HOME
-rm casestudies
+if [ -h casestudies ]
+  then
+    rm casestudies
+fi
 ln -s $IGUANA_CASESTUDIES_HOME/c casestudies
 
+# Make the case study Java code symlink
 cd $IGUANA_HOME/src/main/java/org/iguanatool/
-rm casestudies
+if [ -h casestudies ]
+  then
+    rm casestudies
+fi
 ln -s $IGUANA_CASESTUDIES_HOME/java casestudies
+
+# This is for any specific case study or experimental Java code to be compiled into IGUANA
+# It appears in the "iguana" directory in this repo and maps to org/iguanatool/ext
+# (That is, all code should be in the package org.iguanatool.ext or a subpackage.)
+if [ -h ext ]
+  then
+    rm ext
+fi
+ln -s $IGUANA_CASESTUDIES_HOME/iguana ext
